Замечание 3: Вынесла удаление курьера в @After, теперь курьер удалится в любом случае. Так как есть тесты, в которых курьер не создается, то я вынесла их в отдельный класс.
